Fix deposit type

+++ contracts/PurchaseOfferMarket.sol
@@ -151,7 +151,7 @@ contract PurchaseOfferMarket is Initializable, OwnableUpgradeable, IPurchaseOffe
     function createPurchaseOffer(
         bytes32 _deploymentId,
         uint256 _planTemplateId,
-        uint16 _deposit,
+        uint256 _deposit,
         uint16 _limit,
         uint256 _minimumAcceptHeight,
         uint256 _expireDate
